requirements.txt
pyyaml==6.0.1
pymupdf>=1.22.5
langchain>=0.0.288
sentence-transformers>=2.2.2
transformers>=4.31.0
torch
torchvision
torchaudio
faiss-cpu>=1.7.4
pillow>=10.0.0
numpy>=1.24.3
pandas>=2.0.3
pymilvus>=2.2.8
chromadb>=0.4.13
gradio>=3.40.1
matplotlib>=3.7.2
pytest>=7.4.0
ragas>=0.0.16
datasets>=2.14.4
tqdm>=4.66.1
python-dotenv>=1.0.0
seaborn>=0.12.0
jinja2>=3.1.0